Script.js: add better song/artist parsing and UI tweaks
This commit adds better parsing for song names and artists. It should
play better with arists/songs with periods in them, songs with featured
artists, and the A$AP folk. I also tried to use the Musicxmatch API
but realized that they only provide 30% of the lyrics via the API, qq.
Either way, this should be good enough for now.

I also chose not to use a loop or setInterval to refresh the lyrics.
Figured that refreshing the app is easy enough and didn't want to waste
too much cpu cycles.