No more guessing
Instead of trying to guess what the url for the lyrics will be,
this commit uses Node to spawn a process that runs a Python script.
The Python script uses bs4 to google for the link. This makes lyrics
more reliable. Work to do: it's slow... still misses some lyrics...
and korean music support is necessary.